server/requirements.txt
annotated-types==0.6.0
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
distlib==0.3.8
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
filelock==3.13.3
idna==3.6
instagrapi==2.1.1
pillow==10.3.0
platformdirs==4.2.0
pycryptodomex==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
pygame==2.5.2
PySocks==1.7.1
requests==2.31.0
soupsieve==2.5
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.25.1
